team/teamapp/serializer.py
from rest_framework import serializers
from .models import Member

class MemberSerializer(serializers.ModelSerializer):
    class Meta:
        model = Member
        fields = ['id', 'first', 'last', 'email', 'phone', 'role']